+++ packages/global-package-manager/__tests__/unit/check-validation.js
@@ -12,11 +12,14 @@ jest.mock('../../js/_show-output');
 
 const validationConfig = {
 	required: ['required.md'],
-	ignored: ['ignored'],
 	folders: {
 		folder1: ['scss', 'css'],
 		folder2: ['js', 'json']
 	}
+};
+
+const validationConfigNoFolders = {
+	required: ['required.md']
 };
 
 describe('Check validation', () => {
@@ -82,6 +85,26 @@ describe('Check validation', () => {
 		return expect(
 			checkValidation(validationConfig, 'path/to/global-package', 'failIsTopLevelFile')
 		).rejects.toBeInstanceOf(Error);
+	});
+
+	test('Resolves with any folders when no folder config', () => {
+		const checkValidation = require('../../js/_check-validation');
+		mockfs(MOCK_PACKAGES);
+
+		expect.assertions(1);
+		return expect(
+			checkValidation(validationConfigNoFolders, 'path/to/global-package', 'failIsFolder')
+		).resolves.toEqual();
+	});
+
+	test('Resolves with any files when no folder config', () => {
+		const checkValidation = require('../../js/_check-validation');
+		mockfs(MOCK_PACKAGES);
+
+		expect.assertions(1);
+		return expect(
+			checkValidation(validationConfigNoFolders, 'path/to/global-package', 'failIsFileType')
+		).resolves.toEqual();
 	});
 
 	afterEach(() => {