+++ packages/global-package-manager/__tests__/unit/get-license.js
@@ -6,9 +6,9 @@
 
 const getLicense = require('../../js/_get-license');
 
-jest.mock('package.json', () => ({
+const validationJson = {
 	license: "license-name"
-}));
+};
 
 jest.mock('path/to/package.json', () => ({}), {virtual: true});
 
@@ -16,12 +16,12 @@ describe('Check license', () => {
 	test('Throw error when no license info present in the top level package.json', () => {
 		expect.assertions(1);
 		expect(() => {
-			getLicense('path/to/package.json');
+			getLicense({});
 		}).toThrow();
 	});
 
-    test('Return the licence type from package.json', () => {
-        const license = getLicense('package.json');
+    test('Return the license type from package.json', () => {
+		const license = getLicense(validationJson);
 		expect.assertions(1);
 		expect(license).toEqual('license-name');
 	});