+++ packages/global-package-manager/js/__mocks__/_check-exists.js
@@ -1,0 +1,33 @@
+/**
+ * __mocks__/_check-exists.js
+ * Mocks file/directory checking
+ */
+'use strict';
+
+const fileExistsPath = 'path/to/file.ext';
+const directoryExistsPath = 'path/to/dir';
+
+function fileExists(path) {
+	return new Promise((resolve, reject) => {
+		const localPath = path.replace(`${process.cwd()}/`, '');
+		if (localPath === fileExistsPath) {
+			resolve(localPath);
+		}
+		reject(reject(new Error('invalid file')));
+	});
+}
+
+function directoryExists(path) {
+	return new Promise((resolve, reject) => {
+		const localPath = path.replace(`${process.cwd()}/`, '');
+		if (localPath === directoryExistsPath) {
+			resolve(localPath);
+		}
+		reject(reject(new Error('invalid folder')));
+	});
+}
+
+module.exports = {
+	fileExists,
+	directoryExists
+};