+++ packages/global-package-manager/js/_create.js
@@ -1,4 +1,5 @@
 /**
+ * _create.js
  * Interactive tool to create a new package
  */
 'use strict';
@@ -12,10 +13,10 @@ const tasks = require('./_build-new-package');
 const getPackages = require('./_get-packages');
 const getPackageName = require('./_get-package-name');
 
-module.exports = (validationPath, packageJsonPath, packagesDirectory) => {
-	const config = require(validationPath);
-	const allPackagePaths = getPackages(packagesDirectory);
-	const globalLicense = getLicense(packageJsonPath);
+module.exports = (packageJsonPath, config) => {
+	const packageJson = require(packageJsonPath);
+	const allPackagePaths = getPackages(config.packagesDirectory);
+	const globalLicense = getLicense(packageJson);
 
 	const existingPackages = allPackagePaths.map(currentPath => {
 		return getPackageName(currentPath);
@@ -23,15 +24,15 @@ module.exports = (validationPath, packageJsonPath, packagesDirectory) => {
 
 	inquirer
 		.prompt(
-		prompts(config, existingPackages)
+			prompts(config, existingPackages)
 		)
 		.then(answers => {
 			new Listr(
-				tasks(config, globalLicense, packagesDirectory, answers)
+				tasks(config, globalLicense, config.packagesDirectory, answers)
 			)
-				.run()
-				.catch(err => {
-					console.error(err);
-				});
+			.run()
+			.catch(err => {
+				console.error(err);
+			});
 		});
 };