+++ packages/global-package-manager/js/publish.js
@@ -2,11 +2,26 @@
 
 const path = require('path');
 
-const publishPackages = require(path.resolve(__dirname, './_publish'));
+const publishPackages = require('./_publish');
+const exitScript = require('./_exit-script');
+const exists = require('./_check-exists');
+const configGenerator = require('./_generate-config');
 
 const packageJsonPath = path.resolve(process.cwd(), 'package.json');
-const packageJson = require(packageJsonPath);
 
-publishPackages(
-	path.resolve(process.cwd(), packageJson.packageManager.packagesDirectory)
-);
+configGenerator('package-manager.json')
+	.then(config => {
+		exists.fileExists(packageJsonPath)
+			.then(() => {
+				exists.directoryExists(path.resolve(process.cwd(), config.packagesDirectory))
+					.then(() => {
+						publishPackages(config.packagesDirectory, config.changelog);
+					})
+					.catch(err => {
+						exitScript.displayErr(err);
+					});
+			})
+			.catch(err => {
+				exitScript.displayErr(err);
+			});
+	});