+++ packages/global-package-manager/js/validate.js
@@ -1,14 +1,28 @@
 #! /usr/bin/env node
 
 const path = require('path');
+const argv = require('yargs').argv;
 
-const validatePackages = require(path.resolve(__dirname, './_validate'));
+const validatePackages = require('./_validate');
+const exitScript = require('./_exit-script');
+const exists = require('./_check-exists');
+const configGenerator = require('./_generate-config');
 
 const packageJsonPath = path.resolve(process.cwd(), 'package.json');
-const packageJson = require(packageJsonPath);
 
-validatePackages(
-	path.resolve(process.cwd(), packageJson.packageManager.validationPath),
-	packageJsonPath,
-	path.resolve(process.cwd(), packageJson.packageManager.packagesDirectory)
-);
+configGenerator('package-manager.json')
+	.then(config => {
+		exists.fileExists(packageJsonPath)
+			.then(() => {
+				exists.directoryExists(path.resolve(process.cwd(), config.packagesDirectory))
+					.then(() => {
+						validatePackages(packageJsonPath, config, argv.p);
+					})
+					.catch(err => {
+						exitScript.displayErr(err);
+					});
+			})
+			.catch(err => {
+				exitScript.displayErr(err);
+			});
+	});